+++ R/write_mod.R
@@ -1,7 +1,7 @@
 #' Write a new `mod` file.
 #'
-#' Use `write_mod(file,code)` if you want the `Dynare` file to live in the current working directory.
-#' Use `write_mod(file,code,path)` if you want the `Dynare` file to live in the path different from the current working directory.
+#' Use `write_mod(model,code)` if you want the `Dynare` file to live in the current working directory.
+#' Use `write_mod(model,code,path)` if you want the `Dynare` file to live in the path different from the current working directory.
 #'
 #' @inheritParams run_dynare
 #' @return Set of \code{Dynare} (open-source software for DSGE modelling) outputs
@@ -45,21 +45,21 @@
 #' end;
 #'
 #' stoch_simul;'
-#' file<-"example1" # This is "example1" of the `Dynare` example files
+#' model<-"example1" # This is "example1" of the `Dynare` example files
 #' code<-DynareCodes
-#' write_mod(file,code)
+#' write_mod(model,code)
 #'}
 #' @seealso write_dynare eng_dynare run_model run_dynare
 #' @keywords documentation
 #' @export
-write_mod <- function(file,code,path="") {
-  #file=paste0("DynareR_",file)
+write_mod <- function(model,code,path="") {
+  #model=paste0("DynareR_",model)
 if(path==""){
-  file.create(paste0(file, '.', "mod"))
-  f <-paste0(file, '.', "mod")
+  file.create(paste0(model, '.', "mod"))
+  f <-paste0(model, '.', "mod")
   }else{
-    file.create(paste0(path,"/",file, '.', "mod"))
-  f <-paste0(path,"/",file, '.', "mod")
+    file.create(paste0(path,"/",model, '.', "mod"))
+  f <-paste0(path,"/",model, '.', "mod")
   }
   writeLines(code, f)
 }