+++ demo/agtrend.R
@@ -1,7 +1,7 @@
 # We use "example1" of the Dynare example files to illustrate
 #how to use this function
 
-FileName<-"agtrend"
+ModelName<-"agtrend"
 
 library(DynareR)
 
@@ -184,16 +184,16 @@ check;
 // Plot impulse response functions (Figure 4)
 stoch_simul(order=1) tb_y c_y i_y;'
 
-file<-FileName
+model<-ModelName
 code<-DynareCodes
-run_dynare(file,code)
+run_dynare(model,code)
 
 # You can create an absolute or relative path for the DynareR files.
 # The following writes and run mod file in "DynareR/run_dynare/"  folder
 # relative to the current path.
 
-path=paste0("DynareR","/",file)
+path=paste0("DynareR","/",model)
 
 if(!dir.exists(path)) dir.create(path,recursive = T)
 
-run_dynare(file,code,path)
+run_dynare(model,code,path)