+++ demo/run_dynare.R
@@ -1,7 +1,7 @@
 # We use "example1" of the Dynare example files to illustrate
 #how to use this function
 
-FileName<-"example1"
+modelName<-"example1"
 
 library(DynareR)
 
@@ -48,9 +48,9 @@ var e, u = phi*0.009*0.009;
 end;
 
 stoch_simul;'
-file<-FileName
+model<-modelName
 code<-DynareCodes
-run_dynare(file,code)
+run_dynare(model,code)
 
 # You can create an absolute or relative path for the DynareR files.
 # The following writes and run mod file in "DynareR/run_dynare/"  folder
@@ -60,4 +60,4 @@ path="DynareR/run_dynare"
 
 if(!dir.exists(path)) dir.create(path,recursive = T)
 
-run_dynare(file,code,path)
+run_dynare(model,code,path)