+++ NEWS.md
@@ -3,7 +3,13 @@
 
 What's new?
 
+* Template for R Markdown is created. Go to `file->New File->R Markdown-> From Template->DynareR`.
+
 * DynareR can be used with both base R and R Markdown
 
-* New functions run_dynare run_model, run_models write_dyn and write_mod are created
+* New functions `include_IRF`, `run_dynare`, `run_models`, `write_dyn` and `write_mod` are created
+
+* Embed the graph Inpulse Response Function in R Markdown with `include_IRF`
+
+* Demo files are accessible via demo(package="DynareR")